Move everything related to data exporter in the same folder
