+++ config_collector.go
@@ -1,9 +1,8 @@
 package ffclient
 
 import (
+	"github.com/thomaspoignant/go-feature-flag/exporter"
 	"time"
-
-	"github.com/thomaspoignant/go-feature-flag/internal/dataexporter"
 )
 
 // DataExporter is the configuration of your export target.
@@ -21,5 +20,5 @@ type DataExporter struct {
 
 	// Exporter is the configuration of your exporter.
 	// You can see all available exporter in the exporter package.
-	Exporter dataexporter.Exporter
+	Exporter exporter.Exporter
 }
