+++ feature_flag.go
@@ -3,6 +3,7 @@ package ffclient
 import (
 	"context"
 	"fmt"
+	"github.com/thomaspoignant/go-feature-flag/exporter"
 	"github.com/thomaspoignant/go-feature-flag/internal/dto"
 	"github.com/thomaspoignant/go-feature-flag/retriever"
 	"log"
@@ -12,7 +13,6 @@ import (
 	"github.com/thomaspoignant/go-feature-flag/notifier/logsnotifier"
 
 	"github.com/thomaspoignant/go-feature-flag/internal/cache"
-	"github.com/thomaspoignant/go-feature-flag/internal/dataexporter"
 	"github.com/thomaspoignant/go-feature-flag/internal/fflog"
 )
 
@@ -46,7 +46,7 @@ type GoFeatureFlag struct {
 	cache        cache.Manager
 	config       Config
 	bgUpdater    backgroundUpdater
-	dataExporter *dataexporter.Scheduler
+	dataExporter *exporter.Scheduler
 }
 
 // ff is the default object for go-feature-flag
@@ -92,7 +92,7 @@ func New(config Config) (*GoFeatureFlag, error) {
 
 		if goFF.config.DataExporter.Exporter != nil {
 			// init the data exporter
-			goFF.dataExporter = dataexporter.NewScheduler(goFF.config.Context, goFF.config.DataExporter.FlushInterval,
+			goFF.dataExporter = exporter.NewScheduler(goFF.config.Context, goFF.config.DataExporter.FlushInterval,
 				goFF.config.DataExporter.MaxEventInMemory, goFF.config.DataExporter.Exporter, goFF.config.Logger)
 
 			// we start the daemon only if we have a bulk exporter
